requirements.txt
absl-py==2.3.1
altair==5.5.0
altgraph==0.17.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.2.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dash==3.2.0
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fastjsonschema==2.21.2
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.58.4
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.11.3
kiwisolver==1.4.8
libclang==18.1.1
lxml==6.0.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
narwhals==2.8.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.1
opt_einsum==3.4.0
optree==0.17.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.3.0
parso==0.8.4
patsy==1.0.1
pefile==2023.2.7
pillow==11.2.1
platformdirs==4.3.8
plotly==6.3.1
prompt_toolkit==3.0.51
protobuf==6.33.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
pyzmq==27.0.0
referencing==0.37.0
requests==2.32.4
retrying==1.4.2
rich==14.2.0
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
selenium==4.34.2
setuptools==75.3.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.5
streamlit==1.50.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
wsproto==1.2.0
yellowbrick==1.5
zipp==3.23.0
